[language] kill Fresh/Staged options for building Move stdlib
Nearly all clients use staged, and the ones that use fresh don't need to. This simplifies the code for building the stdlib and lets us delete a fair amount.
