+++ language/compiler/src/main.rs
@@ -17,7 +17,7 @@ use std::{
     io::Write,
     path::{Path, PathBuf},
 };
-use stdlib::{stdlib_modules, StdLibOptions};
+use stdlib::stdlib_modules;
 use structopt::StructOpt;
 use vm::file_format::CompiledModule;
 
@@ -138,7 +138,7 @@ fn main() {
         } else if args.no_stdlib {
             vec![]
         } else {
-            stdlib_modules(StdLibOptions::Staged).to_vec()
+            stdlib_modules().to_vec()
         }
     };
