+++ language/compiler/src/unit_tests/testutils.rs
@@ -8,7 +8,7 @@ use ir_to_bytecode::{
     parser::{parse_module, parse_script},
 };
 use libra_types::{account_address::AccountAddress, vm_error::VMStatus};
-use stdlib::{stdlib_modules, StdLibOptions};
+use stdlib::stdlib_modules;
 use vm::{
     access::ScriptAccess,
     file_format::{CompiledModule, CompiledScript},
@@ -149,7 +149,7 @@ pub fn compile_script_string_with_stdlib(code: &str) -> Result<CompiledScript> {
 }
 
 fn stdlib() -> Vec<CompiledModule> {
-    stdlib_modules(StdLibOptions::Staged)
+    stdlib_modules()
         .iter()
         .map(|m| m.clone().into_inner())
         .collect()
