+++ language/e2e-tests/src/data_store.rs
@@ -15,16 +15,11 @@ use libra_types::{
 use move_vm_state::data_cache::RemoteCache;
 use once_cell::sync::Lazy;
 use std::collections::HashMap;
-use stdlib::StdLibOptions;
 use vm::{errors::*, CompiledModule};
 use vm_genesis::generate_genesis_change_set_for_testing;
 
 /// Dummy genesis ChangeSet for testing
-pub static GENESIS_CHANGE_SET: Lazy<ChangeSet> =
-    Lazy::new(|| generate_genesis_change_set_for_testing(StdLibOptions::Staged));
-
-pub static GENESIS_CHANGE_SET_FRESH: Lazy<ChangeSet> =
-    Lazy::new(|| generate_genesis_change_set_for_testing(StdLibOptions::Fresh));
+pub static GENESIS_CHANGE_SET: Lazy<ChangeSet> = Lazy::new(generate_genesis_change_set_for_testing);
 
 /// An in-memory implementation of [`StateView`] and [`RemoteCache`] for the VM.
 ///
