+++ language/functional-tests/src/evaluator.rs
@@ -31,7 +31,7 @@ use std::{
     str::FromStr,
     time::Duration,
 };
-use stdlib::{stdlib_modules, StdLibOptions};
+use stdlib::stdlib_modules;
 use vm::{
     file_format::{CompiledModule, CompiledScript},
     gas_schedule::{GasAlgebra, MAXIMUM_NUMBER_OF_GAS_UNITS},
@@ -558,12 +558,12 @@ pub fn eval<TComp: Compiler>(
 
     // Set up a fake executor with the genesis block and create the accounts.
     let mut exec = if config.validator_set.payload().is_empty() {
-        FakeExecutor::from_fresh_genesis()
+        FakeExecutor::from_genesis_file()
     } else {
         // use custom validator set. this requires dynamically generating a new genesis tx and
         // is thus more expensive.
         FakeExecutor::custom_genesis(
-            stdlib_modules(StdLibOptions::Fresh).to_vec(),
+            stdlib_modules().to_vec(),
             Some(config.validator_set.clone()),
             VMPublishingOption::Open,
         )
